+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+## Added
+
+## Fixed
+
+## Changed
+
 # 0.3.19-alpha (2024-02-11 / d79ac0c)
 
 Much expanded and improved version, see README for all details. This is