+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 0.9.42 (2024-05-27 / 3b81e73)
 
 ## Changed